last

+++ Infpromo/Models/SearchByUsername.swift
@@ -25,9 +25,9 @@ struct DirectsResponse: Codable {
 }
 
 struct DirectProfileResponse: Codable {
-    let engagementRate: Double?
-    let engagements: Int?
-    let followers: Int?
+    let engagementRate: Float?
+    let engagements: Float?
+    let followers: Float?
     let fullName: String?
     let picture: String?
     let url: String?
